requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pyyaml>=6.0
httpx>=0.28.0
python-telegram-bot>=21.0
